edits to the assembling system
